Add files via upload
We have plotted the scatter map, to plot a scatterplot the face_positions list is initialized to keep track of the position of detected faces. Each time a face is detected, a rectangle is drawn around it and its position are added to the list. After the monitoring is finished, x and y coordinates of the face positions are extracted using the zip function, and a scatter plot is then produced using the plt.scatter function from the Matplotlib toolkit. This scatter map can be useful for identifying hotspots